dev: 주문 API 기능 완료 및 테스트 코드 작성 - 컨트롤러 테스트 - 서비스 & 레포지토리 테스트

+++ src/main/java/kr/hhplus/be/server/domain/order/application/dto/OrderCreateResponse.java
@@ -1,20 +1,21 @@
 package kr.hhplus.be.server.domain.order.application.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderCreateResponse(
         String orderId,
         List<OrderResItem> items,
-        double totalPrice,
+        BigDecimal totalPrice,
         String orderDt,
         String orderStatus
     ) {
 
     public record OrderResItem(
             Long productLineId,
-            Double linePrice,
+            BigDecimal linePrice,
             String couponYN,
-            Double discountPrice,
+            BigDecimal discountPrice,
             int quantity,
             String orderStatus
     ) {}
